check

+++ training/behavior_llm_sft.py
@@ -66,8 +66,8 @@ EOS_TOKEN = tokenizer.eos_token
 if load_in_8bit:
     quant_config = BitsAndBytesConfig(
     load_in_8bit=True,
-    # bnb_8bit_compute_dtype=torch.bfloat16,
-    bnb_8bit_compute_dtype=torch.float32,   # <- critical change
+    bnb_8bit_compute_dtype=torch.bfloat16,
+    #bnb_8bit_compute_dtype=torch.float32,   # <- critical change
     llm_int8_threshold=6.0,            # optional tuning threshold
     # llm_int8_has_fp16_weight=False,    # optional
     )
